Add files via upload
Here is a code , to perform bruteforce on an encrypted zip file which is obviously personal (in this scenario). It happens sometimes that we set a password and then forget and to remember our passwords we often note them in a diary or a text file (here the text file is used) and we often have more than 20 passwords at a same place.
So rather checking it manually, why not to try something interesting...
This is all about learning  something new. 
So keep learning...